+++ ex-5.py
@@ -1,7 +1,7 @@
 somaDosPesos = 0
 
 for numeroCaixa in range(1, 26):
-    pesoCaixa = int(input('Digite o peso da caixa ' + str(numeroCaixa) + ': ' ))
+    pesoCaixa = float(input('Digite o peso da caixa ' + str(numeroCaixa) + ': ' ))
     somaDosPesos += pesoCaixa
     
 print('A soma dos pesos de todas as caixas é:', somaDosPesos)
